% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fluide.R
\name{fluid_properties}
\alias{fluid_properties}
\title{Physical properties of standard fluids as a function of temperature}
\usage{
fluid_properties(fluid = "water", T = 20, g = 9.81)
}
\arguments{
\item{fluid}{Character specifying the desired fluid. Options:
\describe{
\item{"water"}{Pure water}
\item{"air"}{Dry air}
\item{"oil"}{SAE 30 oil}
}
Default: "water"}

\item{T}{Fluid temperature in degrees Celsius (?C).
Recommended ranges:
\itemize{
\item Water: 0 to 100?C
\item Air: -20 to 100?C
\item Oil: 0 to 100?C
}
Default: 20}

\item{g}{Gravitational acceleration in m/s2. Default: 9.81}
}
\value{
A list of class "fluid_properties" containing:
\describe{
\item{name}{Full name}
\item{rho}{Density (kg/m3)}
\item{mu}{Dynamic viscosity (Pa.s)}
\item{nu}{Kinematic viscosity (m2/s)}
\item{gamma}{Specific weight (N/m3)}
\item{T}{Temperature (?C)}
\item{fluid_type}{Fluid type}
}
}
\description{
Returns the main physical properties of water, air, and oil
in fluid mechanics, with correction for the specified temperature.
}
\details{
This function provides the properties of three basic fluids:
\itemize{
\item \code{name} : Full name of the fluid
\item \code{rho} : Density in kg/m3
\item \code{mu} : Dynamic viscosity in Pa.s
\item \code{nu} : Kinematic viscosity in m2/s (nu/?)
\item \code{gamma} : Specific weight in N/m3 (rho ? g)
\item \code{T} : Actual temperature used in ?C
}
}
\examples{
Properties of water at different temperatures
fluid_properties("water", 20)
fluid_properties("water", 50)
fluid_properties("water", 100)

Properties of air
fluid_properties("air", 0)
fluid_properties("air", 25)

Properties of oil
fluid_properties("oil", 20)
fluid_properties("oil", 40)

Use in a calculation
props <- fluid_properties("water", 80)
cat("Density of water at 80?C:", props$rho, "kg/m3\n")
cat("Viscosity:", props$mu, "Pa.s\n")

}
\references{
White, F. M. (2011). \emph{Fluid Mechanics} (7th ed.). McGraw-Hill.
}
\seealso{
\code{\link{bernoulli_standard}}
}
\concept{fluid_properties_functions}
\keyword{density}
\keyword{fluid}
\keyword{properties}
\keyword{temperature}
\keyword{viscosity}
