% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fluide.R
\name{print.fluid_properties}
\alias{print.fluid_properties}
\title{Display fluid properties}
\usage{
\method{print}{fluid_properties}(x, ...)
}
\arguments{
\item{x}{fluid_properties object}

\item{...}{Additional arguments}
}
\description{
Print method for fluid_properties objects
}
