% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fluide.R
\name{print.fluid_properties}
\alias{print.fluid_properties}
\title{Affichage des propriétés d'un fluide}
\usage{
\method{print}{fluid_properties}(x, ...)
}
\arguments{
\item{x}{Objet fluid_properties}

\item{...}{Arguments supplémentaires}
}
\description{
Méthode print pour les objets fluid_properties
}
